requirements.txt
absl-py==1.3.0
# aniso8601==9.0.1
# # apache-beam==2.34.0
# astunparse==1.6.3
# attrs==21.2.0
# avro-python3==1.9.2.1
# blas==1.0
# blosc==1.21.0
# brotli==1.0.9
# brunsli==0.1 
# bzip2==1.0.8
# c-ares==1.18.1
# ca-certificates==2020.10.14
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
# cfitsio==3.470
# charls==2.2.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
cython==0.29.25
cytoolz==0.11.0
# dask-core==2021.10.0
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
et_xmlfile==1.1.0
fastavro==1.4.7
flask==2.0.2
flask-cors==3.0.10
flask-restx==0.5.1
flatbuffers==22.10.26
# fonttools==4.28.3
# freetype==2.11.0
fsspec==2021.10.1
funcy==1.17
future==0.18.2
gast==0.4.0
# giflib==5.2.1
gin-config==0.5.0
# google-api-core==2.10.2
# google-api-python-client==2.65.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
# googleapis-common-protos==1.56.4
grpcio==1.34.1
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
# imagecodecs==2021.8.26
imageio==2.9.0
importlib-metadata==4.8.2
imutils==0.5.4
intel-openmp==2021.4.0
itsdangerous==2.0.1
jinja2==3.0.3
joblib==1.1.0 
# jsonschema==4.7.2
# jxrlib==1.1 
kaggle==1.5.12
keras==2.7.0
keras-nightly==2.11.0.dev2022092207
# keras-nightly==2.12.0.dev2022110708
keras-preprocessing==1.1.2
# kiwisolver==1.3.2
# krb5==1.19.2  
# lcms2==2.12  
# ld_impl_linux-64==2.35.1  
# lerc==3.0  
# libaec==1.0.4  
libclang==14.0.6
# libcurl==7.80.0 
# libdeflate==1.8  
# libedit==3.1.20210910
# libev==4.33 
# libffi==3.3  
# libgcc-ng==9.3.0  
# libgfortran-ng==7.5.0  
# libgfortran4==7.5.0  
# libgomp==9.3.0  
# libnghttp2==1.46.0
# libpng==1.6.37 
# libssh2==1.9.0  
# libstdcxx-ng==9.3.0  
# libtiff==4.2.0  
# libwebp==1.2.0  
# libwebp-base==1.2.0  
# libzopfli==1.0.3 
# llvmlite==0.38.0
# locket==0.2.1  
# lvis==0.5.3
# lxml==4.7.0
# lz4-c==1.9.3
markdown==3.3.6
markupsafe==2.0.1
matplotlib==3.3.4
# matplotlib-base==3.5.0
mkl==2021.4.0
# mkl-service==2.4.0
# mkl_fft==1.3.1
# mkl_random==1.2.2
# munkres==1.1.4
# ncurses==6.3
# networkx==2.6.3
nltk==3.6.5
# numba==0.55.0
numpy==1.19.5
# numpy-base==1.21.2
oauth2client==4.1.3
oauthlib==3.1.1
# object-detection==0.1
olefile==0.46
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
# openjpeg==2.4.0 
openpyxl==3.0.9
# openssl==1.1.1m
opt-einsum==3.3.0
orjson==3.6.1
packaging==21.3
pandas==1.1.5
partd==1.2.0  
pillow==8.4.0
pip==21.2.4
portalocker==2.3.2
promise==2.3
protobuf==3.19.6
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycocotools==2.0.3
pycparser==2.21
pydot==1.4.2
pymongo==3.12.3
pyparsing==2.4.7
pyrsistent==0.18.0
pysnic==1.0.4
pytesseract==0.3.9
# python==3.9.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pywavelets==1.1.1  
pyyaml==6.0
# readline==8.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
# rsa==4.8
# sacrebleu==2.0.0
# scikit-image==0.18.3  
# scikit-learn==1.0.1
# scipy==1.7.3  
# sentencepiece==0.1.96
# seqeval==1.2.2
setuptools==59.6.0
six==1.15.0
# snappy==1.1.8 
# sounddevice==0.4.5
# sqlite==3.36.0  
tabulate==0.8.9
tb-nightly==2.11.0a20220816
#tb-nightly==2.11.0a20221107 
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
# tensorflow==2.7.0
tensorflow==2.6.2
# tensorflow-addons==0.15.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.12.0
# tensorflow-io==0.22.0
tensorflow-io==0.21.0
# tensorflow-io-gcs-filesystem==0.27.0           
# tensorflow-metadata==1.5.0
tensorflow-metadata==1.2.0
tensorflow-model-optimization==0.7.0
# tensorflow-text==2.7.3
termcolor==1.1.0
text-unidecode==1.3
# tf-estimator-nightly==2.12.0.dev2022110709
# tf-models-official==2.7.0
# tf-nightly==2.12.0.dev20221107
tf-slim==1.1.0
# tflite-support==0.4.3
tflite-support==0.3.0
threadpoolctl==3.0.0
# tifffile==2021.7.2  
# tk==8.6.11  
# toolz==0.11.2
torch==1.10.2  
# torch==1.12.0
torchvision==0.11.2
# torchvision==0.13.0 
tqdm==4.62.3 
typeguard==2.13.3
typing-extensions==3.7.4.3
# tzdata==2021e  
uritemplate==4.1.1
urllib3==1.26.7
werkzeug==2.0.2
# wheel==0.37.0
wrapt==1.12.1
# xz==5.2.5
# yaml==0.2.5 
# zlib==1.2.11  

